fix(client-config.ts): set useCdn property to false to disable CDN usage
The useCdn property in the client configuration file was set to true, which means that the Sanity client was using the Content Delivery Network (CDN) to fetch data. However, this was causing some caching issues and outdated data being displayed. To address this, the useCdn property has been changed to false, disabling the CDN usage and ensuring that the latest data is always fetched from the Sanity API.